relay: limit NOTICE queue size, don't store all OK
Adds length check (max 4bytes per char) to the message
we receive so we don't accept huge messages.

Stops storing all `NOTICE` message in a queue forever by limiting the
maxsize of the queue to 100 (they aren't used for anything right now).

Stops storing all `OK` messages in a queue, instead replaces the queue
with a dict of futures. This allows to assign the `OK` response
to the according sent event id instead of just accepting any `OK`
as response, even if it may was meant for another event.
Also stops storing `OK` responses if a response is not required by the sender.