manager: don't stop only_stored sub on first EOSE
Don't stop a subscription with only_stored=True on the first EOSE
message we get from any relay. Wait until we either got a EOSE from all
connected relays or until Manager.EOSE_TIMEOUT_SEC.
Currently it was a race and we might miss events as we instantly stop
the subscription if any of our connected relays sends EOSE, not waiting
if any other relay might has the events we requested.